+++ lib/src/main/java/uk/gov/di/ipv/cri/address/library/exception/ClientConfigurationException.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.cri.address.library.exceptions;
+package uk.gov.di.ipv.cri.address.library.exception;
 
 public class ClientConfigurationException extends Exception {
     public ClientConfigurationException(Exception e) {
